requirements.txt
amqp==5.0.9
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.6.0
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coverage==6.2
Deprecated==1.2.13
Django==3.2.8
django-cpf==0.1.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-filter==21.1
djangorestframework==3.13.0
djangorestframework-simplejwt==5.0.0
Faker==9.9.0
iniconfig==1.1.1
kombu==5.2.3
mixer==7.2.0
packaging==21.3
Pillow==8.3.2
pluggy==1.0.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.1
py==1.11.0
pycodestyle==2.8.0
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.5.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
